+++ examples/weather/weatherbot.rb
@@ -5,8 +5,8 @@ module WeatherBot
   end
 
   class Weather < SlackRubyBot::Commands::Base
-    match(/^How is the weather in (?<location>\w*)\?$/i) do |data, match|
-      send_message data.channel, "The weather in #{match[:location]} is nice."
+    match(/^How is the weather in (?<location>\w*)\?$/i) do |client, data, match|
+      send_message client, data.channel, "The weather in #{match[:location]} is nice."
     end
   end
 end