+++ lib/slack-ruby-bot/commands/about.rb
@@ -4,8 +4,8 @@ module SlackRubyBot
       command 'about'
       match(/^(?<bot>\w*)$/)
 
-      def self.call(data, _match)
-        send_message_with_gif data.channel, SlackRubyBot::ABOUT, 'selfie'
+      def self.call(client, data, _match)
+        send_message_with_gif client, data.channel, SlackRubyBot::ABOUT, 'selfie'
       end
     end
   end