+++ lib/slack-ruby-bot/commands/help.rb
@@ -1,8 +1,8 @@
 module SlackRubyBot
   module Commands
     class Help < Base
-      def self.call(data, _match)
-        send_message_with_gif data.channel, 'See https://github.com/dblock/slack-ruby-bot, please.', 'help'
+      def self.call(client, data, _match)
+        send_message_with_gif client, data.channel, 'See https://github.com/dblock/slack-ruby-bot, please.', 'help'
       end
     end
   end