+++ lib/slack-ruby-bot/commands/hi.rb
@@ -1,8 +1,8 @@
 module SlackRubyBot
   module Commands
     class Hi < Base
-      def self.call(data, _match)
-        send_message_with_gif data.channel, "Hi <@#{data.user}>!", 'hi'
+      def self.call(client, data, _match)
+        send_message_with_gif client, data.channel, "Hi <@#{data.user}>!", 'hi'
       end
     end
   end