+++ lib/slack-ruby-bot/commands/unknown.rb
@@ -3,8 +3,8 @@ module SlackRubyBot
     class Unknown < Base
       match(/^(?<bot>\w*)[\s]*(?<expression>.*)$/)
 
-      def self.call(data, _match)
-        send_message_with_gif data.channel, "Sorry <@#{data.user}>, I don't understand that command!", 'idiot'
+      def self.call(client, data, _match)
+        send_message_with_gif client, data.channel, "Sorry <@#{data.user}>, I don't understand that command!", 'idiot'
       end
     end
   end